Refactor: rename dbconnector in db
